fix: bug in datatype factorial_function

+++ 7_Branching - Looping - Function/praktikum/factorial_function.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-int factorial(int n) {
+double factorial(int n) {
   if (n == 0) {
     return 1;
   } else {
